long ID

+++ logic/src/main/java/application/Repository/ProductRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
 
-    Product findById(int id);
+    Product findById(long id);
 
 }
